Refactor CLI Composition
