+++ docs/en/architecture.md
@@ -205,7 +205,7 @@ Specifics:
 
 ### 5. `apps/cli` — Entry point and composition root
 
-`analyze-waste.command.ts` is the only place where concrete implementations are instantiated: it loads the config file, builds the price table, the policies (with config + CLI parameters) and 9 of the 11 scanners, injects them into the use case and hands the result to the formatters. The other two, `AwsEc2UnderutilizedScanner` and `AwsRdsUnderutilizedScanner`, are registered only when `--live-pricing` is set: their savings estimate needs a per-instance-type/class price that the static table doesn't carry, so without live pricing there is nothing reliable to report and the scanners are left out rather than registered with a zero estimate. The four formatters (console table, PDF, JSON, Markdown) share the `resource-presenters.ts` registry, typed `Record<ResourceKind, ResourcePresenter<…>>`: forgetting the presenter for a new kind is a compile error. The output format is selected by `--format` (`table` | `json` | `markdown`); `markdown` targets CI / PR comments.
+`analyze-waste.command.ts` loads the config file, resolves CLI options (regions, min-age, account ID) and orchestrates the run; it delegates the actual instantiation of concrete implementations to `analyze-waste.composition.ts` through the injectable `AnalyzeDeps.createAnalysis` seam (the same seam `analyze-waste.command.spec.ts` fakes to test without AWS). `analyze-waste.composition.ts` is the only place where concrete implementations are instantiated: it builds the price table, the policies (with config + CLI parameters) and 9 of the 11 scanners, injecting them into the use case. The other two, `AwsEc2UnderutilizedScanner` and `AwsRdsUnderutilizedScanner`, are registered only when `--live-pricing` is set: their savings estimate needs a per-instance-type/class price that the static table doesn't carry, so without live pricing there is nothing reliable to report and the scanners are left out rather than registered with a zero estimate. Back in `analyze-waste.command.ts`, the result is handed to the formatters. The four formatters (console table, PDF, JSON, Markdown) share the `resource-presenters.ts` registry, typed `Record<ResourceKind, ResourcePresenter<…>>`: forgetting the presenter for a new kind is a compile error. The output format is selected by `--format` (`table` | `json` | `markdown`); `markdown` targets CI / PR comments.
 
 ---
