+++ docs/it/architettura.md
@@ -206,7 +206,7 @@ Particolarità:
 
 ### 5. `apps/cli` — Entry point e composition root
 
-`analyze-waste.command.ts` è l'unico punto in cui le implementazioni concrete vengono istanziate: carica il file di config, costruisce il listino prezzi, le policy (con i parametri da config + CLI) e 9 degli 11 scanner, li inietta nel use case e passa il risultato ai formatter. Gli altri due, `AwsEc2UnderutilizedScanner` e `AwsRdsUnderutilizedScanner`, vengono registrati solo se `--live-pricing` è attivo: la loro stima di risparmio richiede un prezzo per instance type/classe che il listino statico non contiene, quindi senza prezzi live non c'è nulla di affidabile da riportare e gli scanner vengono esclusi piuttosto che registrati con una stima a zero. I quattro formatter (tabella console, PDF, JSON, Markdown) condividono il registry `resource-presenters.ts`, tipizzato `Record<ResourceKind, ResourcePresenter<…>>`: dimenticare il presenter di un nuovo kind è un errore di compilazione. Il formato di output si sceglie con `--format` (`table` | `json` | `markdown`); `markdown` è pensato per CI / commenti PR.
+`analyze-waste.command.ts` carica il file di config, risolve le opzioni CLI (regioni, min-age, account ID) e orchestra l'esecuzione; delega l'istanziazione effettiva delle implementazioni concrete a `analyze-waste.composition.ts` tramite il seam iniettabile `AnalyzeDeps.createAnalysis` (lo stesso seam che `analyze-waste.command.spec.ts` finge per testare senza AWS). `analyze-waste.composition.ts` è l'unico punto in cui le implementazioni concrete vengono istanziate: costruisce il listino prezzi, le policy (con i parametri da config + CLI) e 9 degli 11 scanner, iniettandoli nel use case. Gli altri due, `AwsEc2UnderutilizedScanner` e `AwsRdsUnderutilizedScanner`, vengono registrati solo se `--live-pricing` è attivo: la loro stima di risparmio richiede un prezzo per instance type/classe che il listino statico non contiene, quindi senza prezzi live non c'è nulla di affidabile da riportare e gli scanner vengono esclusi piuttosto che registrati con una stima a zero. Tornati in `analyze-waste.command.ts`, il risultato passa ai formatter. I quattro formatter (tabella console, PDF, JSON, Markdown) condividono il registry `resource-presenters.ts`, tipizzato `Record<ResourceKind, ResourcePresenter<…>>`: dimenticare il presenter di un nuovo kind è un errore di compilazione. Il formato di output si sceglie con `--format` (`table` | `json` | `markdown`); `markdown` è pensato per CI / commenti PR.
 
 ---
